requirements.txt
viam-sdk>=0.21

numpy <2
torch==2.0
torchvision == 0.14
opencv-python
pillow

scipy


tabulate
# ##from torchreid requirements
# numpy
# Cython
# h5py
# Pillow
# six
# scipy
# opencv-python
# matplotlib
# tb-nightly
# future
# yacs
# gdown
# flake8
# yapf
# isort==4.3.21
# imageio
# chardet

## for testing and linting
pytest
pytest-asyncio
pylint

## for building
pyinstaller